requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
tensorflow-cpu==2.13.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
Pillow==10.1.0
reportlab==4.1.0
matplotlib==3.7.1
gunicorn==21.2.0
